Remove soft merge mode.
Soft merging has been removed, as there are some footguns that cannot
easily be fixed.

soft mode would never merge, if the target is non-optional.
Due to aliasing however, we can never be actually sure that any type is
not optional, since it's impractically to do full type resolution in the
proc macro state!